Se agrego el rebote del balon al cochar con los bordes de pantallas

+++ ClientGame/src/models/Ball.java
@@ -2,27 +2,27 @@ package models;
 
 public class Ball {
 
-	private int posX;
-	private int posY;
+	private float posX;
+	private float posY;
 	
-	public Ball(int posX, int posY) {
+	public Ball(float posX, float posY) {
 		this.posX = posX;
 		this.posY = posY;
 	}
 
-	public int getPosX() {
+	public float getPosX() {
 		return posX;
 	}
 
-	public void setPosX(int posX) {
+	public void setPosX(float posX) {
 		this.posX = posX;
 	}
 
-	public int getPosY() {
+	public float getPosY() {
 		return posY;
 	}
 
-	public void setPosY(int posY) {
+	public void setPosY(float posY) {
 		this.posY = posY;
 	}
 }
